refactor: duplicate code on employee and service page
