--- frontend/components/pathBuilder.js
@@ -1,5 +1,0 @@
-const BuildPath = (path) => {
-    return 'http://localhost:5000/api/v1/' + path;
-};
-
-export default BuildPath;
