+++ frontend/pages/employee/index.js
@@ -1,5 +1,5 @@
 import EmployeeComponent from "../../components/employee";
-
+import groupService from "../../utils/groupService";
 import {useState} from "react";
 import _groupBy from "lodash/groupBy";
 import _cloneDeep from "lodash/cloneDeep";
@@ -8,25 +8,7 @@ import {http} from "../../utils/http";
 export async function getServerSideProps(context) {
     const employeeList = await http(`/api/v1/employees`);
     const serviceListResponse = await http(`/api/v1/services`);
-    const serviceArray = _groupBy(serviceListResponse, 'service_code')
-    let serviceList = [];
-    let item, durationPriceList, durationPriceItem;
-    for (const serviceCode in serviceArray) {
-        item = _cloneDeep(serviceArray[serviceCode][0]);
-        item.service_code = serviceCode;
-        delete item.duration;
-        delete item.price;
-        durationPriceList = [];
-        for (let insideItem of serviceArray[serviceCode]) {
-            durationPriceItem = {};
-            durationPriceItem.id = insideItem.id;
-            durationPriceItem.duration = (insideItem.duration * 1 / 3600000).toFixed(2);
-            durationPriceItem.price = insideItem.price;
-            durationPriceList.push(durationPriceItem);
-        }
-        item.durations_prices = durationPriceList;
-        serviceList.push(item);
-    }
+    const serviceList = groupService(serviceListResponse);
     return {
         props: {serviceList: serviceList, employeesList: employeeList},
     };
