+++ frontend/pages/service/index.js
@@ -1,8 +1,7 @@
 import ServiceComponent from '../../components/service';
 import {useEffect, useState} from 'react';
 import {http} from "../../utils/http";
-import _groupBy from "lodash/groupBy";
-import _cloneDeep from "lodash/cloneDeep";
+import groupService from "../../utils/groupService";
 
 export async function getServerSideProps(context) {
     const employeeList = await http(`/api/v1/employees`);
@@ -15,28 +14,6 @@ const Service = ({employeeList}) => {
     const [serviceListData, setServiceListData] = useState([]);
     const [refresh, setRefresh] = useState(false);
     const [loading, setLoading] = useState(undefined);
-    const groupService = (serviceListResponse) => {
-        const serviceArray = _groupBy(serviceListResponse, 'service_code')
-        const serviceList = [];
-        let item, durationPriceList, durationPriceItem;
-        for (const serviceCode in serviceArray) {
-            item = _cloneDeep(serviceArray[serviceCode][0]);
-            item.service_code = serviceCode;
-            delete item.duration;
-            delete item.price;
-            durationPriceList = [];
-            for (let insideItem of serviceArray[serviceCode]) {
-                durationPriceItem = {};
-                durationPriceItem.id = insideItem.id;
-                durationPriceItem.duration = (insideItem.duration * 1 / 3600000).toFixed(2);
-                durationPriceItem.price = insideItem.price;
-                durationPriceList.push(durationPriceItem);
-            }
-            item.durations_prices = durationPriceList;
-            serviceList.push(item);
-        }
-        return serviceList;
-    }
 
     const getServiceList = async () => {
         const serviceListResponse = await http(`/api/v1/services`);
